🩹 Made Bot only add new users on startup.
Made bot add users that joined while bot was offline to user database.
Also tried to catch aiohttp exception that occures when bot disconnects and discord is not available.
This could spam the log if the connection is unstable from time to time, so I tried to mittigate that.
Did not succed tho.