🗃️ Reset alembic migrations.
Had to reset Alemic Migration because of misstakes earlier...
This is not good, but was long neccesary!
First migration were not autogenerated, so they were empty and If I wanted to fix them manually, I would need to implement the upgrades and downgrades manually.
This would not be practical, so I opted to delete all migrations and start fresh!